update feature course management and feature other

+++ spring-microservices/course-service/src/main/java/com/hoangphihiep/dto/response/AssignmentGradingResponse.java
@@ -27,7 +27,7 @@ public class AssignmentGradingResponse {
 
     private Integer pendingAssignments;
 
-    private Integer averageScore;
+    private Double averageScore;
 
     private List<AssignmentSubmissionResponse> latestSubmissions;
 }
